fix(wiki): rework how an app is deployed so we can deploy new versions easier
Basically, all parameters of the template are stored in Nomad Variables.
They can easily be collected and used to re-render the template, to
make a new deployment.

A simple "pulumi refresh" will fetch the active version into the
pulumi stack, allowing us to redeploy the same version from pulumi
too.