feat(nomad-core): nginx proxy to load balance services to all Nomad hosts
This proxy automatically reconfigures if new services come online.
Every service needs to have a tag "port=" with an unique port behind
it. Nginx will listen on this port, and distribute the requests
over the services running.

Nginx runs as "system", which means that it is reachable on every
Nomad node.