fix(aws-core): switch to a single interface for the Nomad cluster
Having two interfaces means delays on creating an outbound IPv6
connection, as it first tries the interface with the prefix.

Normally you can have an interface with both a prefix and an
address, but for some unknown reason ASG doesn't want to start
a launch template that is configured as such.

So, on start-up, we pick an IP from the prefix ourselves and put
that on the interface. This does mean it needs to be done as
early as possible.