feat(app/wiki): use an R2 bucket for additional caching
This means that the wiki is served first from cache, then from R2,
and if those all fail, from the origin. And of course it checks
with origin on every request if it is still up-to-date, but this
now returns in most cases a 304, which are far cheaper than the
big images that are on the wiki.