chore(app/wiki): various of fixes for how objects are cached
Most noteworthy is that Cloudflare removed "If-None-Match" from
a fetch request if it wants to fill its own cache, even if the
resulting object doesn't allow that. By making it explicit we
don't want to cache, this behaviour is prevented. In other words:

The backend checks if the content was stale only works if we tell
Cloudflare to not remove the "If-None-Match" header.